module-bluetooth-policy: Remember last used profile.
When switching profile to HSP due to the auto_switch parameter
being set, make module-bluetooth-policy remember the current profile
in use by the card. In this way, when the reverting to the original
profile, the card won't fallback to the default A2DP profile, but
will go back to whatever profile was being used at the time of switching
(e.g. LDAC).